requirements.txt
streamlit
pandas
snowflake-snowpark-python
snowflake-ml-python
matplotlib
